Remove config.ts
