--- packages/server/lib/modes/init/config.ts
@@ -1,10 +1,0 @@
-export interface Config {
-  projectRoot: string
-  fixturesFolder?: string | false
-  integrationFolder?: string
-  pluginsFile?: string | false
-  supportFile?: string | false
-  screenshotsFolder?: string
-  videosFolder?: string
-  video?: boolean
-}
